src/requirements.txt
tweepy >= 4.8.0
python-dotenv >= 0.20.0